PERSONAL/STUDY/2024/GUI-CHALLENGES/ai/ai-lab/RAG/code/requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
llama-index==0.10.0
openai
openpyxl
psycopg2-binary

python-dotenv

SQLAlchemy